Refactor the handling of Modal
Simplify Modal component as it does a lot of complicated useEffects. The problem is that the conditional statement for Modal in Gallery is simply removing/displaying the component instantly.

Instead of that, use a useState if we want to display the Modal

Also, make sure the Modal will accept (look pretty with):
- different image sizes and orientation
- transparent backgrounds